requirements.txt
numpy>=1.19.5
pandas==1.3.5
requests-html>=0.10.0
ta>=0.7.0
tqdm>=4.55.1
yfinance==0.1.67
openpyxl>=3.0.6
slackclient>=2.9.4